refactor: move httpserver folder
