+++ .github/workflows/publish-dev-assets.yml
@@ -79,7 +79,7 @@ jobs:
       - name: docker build ratify base
         run: |
           docker buildx create --use         
-          docker buildx build -f ./httpserver/Dockerfile \
+          docker buildx build -f ./Dockerfile \
             --attest type=sbom \
             --attest type=provenance,mode=max \
             --platform linux/amd64,linux/arm64,linux/arm/v7 \
@@ -91,7 +91,7 @@ jobs:
       - name: docker build ratify with plugin
         run: |
           docker buildx create --use
-          docker buildx build -f ./httpserver/Dockerfile \
+          docker buildx build -f ./Dockerfile \
             --attest type=sbom \
             --attest type=provenance,mode=max \
             --platform linux/amd64,linux/arm64,linux/arm/v7 \
