+++ server/httpserver/types.go
