+++ server/ratifymain/main.go
@@ -17,15 +17,13 @@ package ratifymain
 
 import (
 	"context"
-	"flag"
 	"fmt"
-	"time"
 
 	"github.com/ratify-project/ratify/config"
-	"github.com/ratify-project/ratify/httpserver"
 	"github.com/ratify-project/ratify/internal/logger"
 	"github.com/ratify-project/ratify/pkg/cache"
-	"github.com/ratify-project/ratify/pkg/manager"
+	"github.com/ratify-project/ratify/server/httpserver"
+	"github.com/ratify-project/ratify/server/manager"
 	"github.com/sirupsen/logrus"
 
 	_ "github.com/ratify-project/ratify/pkg/cache/dapr"                  // register dapr cache
@@ -37,52 +35,11 @@ import (
 	_ "github.com/ratify-project/ratify/pkg/verifier/notation"           // register notation verifier
 )
 
-// options defines the command line options to start the ratify server.
-type options struct {
-	configFilePath    string
-	httpServerAddress string
-	certDirectory     string
-	caCertFile        string
-	enableCrdManager  bool
-	cacheEnabled      bool
-	cacheType         string
-	cacheName         string
-	cacheSize         int
-	cacheTTL          time.Duration
-	metricsEnabled    bool
-	metricsType       string
-	metricsPort       int
-	healthPort        string
-}
-
 func Main() {
 	opts := parse()
 	if err := startRatify(opts); err != nil {
 		logrus.Errorf("Error starting Ratify: %v", err)
 	}
-}
-
-// parse parses the command line arguments and returns the options.
-func parse() *options {
-	opts := options{}
-	flag.StringVar(&opts.httpServerAddress, "http", "", "HTTP Address")
-	flag.StringVar(&opts.configFilePath, "config", "", "Config File Path")
-	flag.StringVar(&opts.certDirectory, "cert-dir", "", "Path to ratify certs")
-	flag.StringVar(&opts.caCertFile, "ca-cert-file", "", "Path to CA cert file")
-	flag.BoolVar(&opts.enableCrdManager, "enable-crd-manager", false, "Start crd manager if enabled (default: false)")
-	flag.BoolVar(&opts.cacheEnabled, "cache-enabled", false, "Enable cache if enabled (default: false)")
-	flag.StringVar(&opts.cacheType, "cache-type", cache.DefaultCacheType, fmt.Sprintf("Cache type to use (default: %s)", cache.DefaultCacheType))
-	flag.StringVar(&opts.cacheName, "cache-name", cache.DefaultCacheName, fmt.Sprintf("Cache implementation name to use (default: %s)", cache.DefaultCacheName))
-	flag.IntVar(&opts.cacheSize, "cache-size", cache.DefaultCacheSize, fmt.Sprintf("Cache max size to use in MB (default: %d)", cache.DefaultCacheSize))
-	flag.DurationVar(&opts.cacheTTL, "cache-ttl", cache.DefaultCacheTTL, fmt.Sprintf("Cache TTL for the verifier http server (default: %fs)", cache.DefaultCacheTTL.Seconds()))
-	flag.BoolVar(&opts.metricsEnabled, "metrics-enabled", false, "Enable metrics exporter if enabled (default: false)")
-	flag.StringVar(&opts.metricsType, "metrics-type", httpserver.DefaultMetricsType, fmt.Sprintf("Metrics exporter type to use (default: %s)", httpserver.DefaultMetricsType))
-	flag.IntVar(&opts.metricsPort, "metrics-port", httpserver.DefaultMetricsPort, fmt.Sprintf("Metrics exporter port to use (default: %d)", httpserver.DefaultMetricsPort))
-	flag.StringVar(&opts.healthPort, "health-port", httpserver.DefaultHealthPort, fmt.Sprintf("Health port to use (default: %s)", httpserver.DefaultHealthPort))
-	flag.Parse()
-
-	logrus.Infof("Starting Ratify: %+v", opts)
-	return &opts
 }
 
 // startRatify starts the ratify server.
